added apiURL as a variable
As resolution to issue#2, the apiURL variable was created containing the first section of a commonly used API.

Lines of code where this url was referenced were updated to refer to the variable instead of the whole url.

Additionally, the order of surge vs localhost clientID's and redirectURLs were modified for easier reading.